refactor: Use page parts structure
