+++ docs/src/components/page-parts/directory-structure/app-vite/DirectoryStructure.vue
@@ -1,0 +1,22 @@
+<template lang="pug">
+base-directory-structure(:structure="structure")
+</template>
+
+<script>
+import BaseDirectoryStructure from '../BaseDirectoryStructure.vue'
+import structure from './structure'
+
+export default {
+  name: 'AppWebpackDirectoryStructure',
+
+  components: {
+    BaseDirectoryStructure
+  },
+
+  setup () {
+    return {
+      structure
+    }
+  }
+}
+</script>
