+++ docs/src/components/page-parts/directory-structure/app-vite/structure.js
