+++ docs/src/components/page-parts/directory-structure/app-webpack/structure.js
