+++ docs/src/pages/quasar-cli-vite/directory-structure.md
@@ -1,6 +1,8 @@
 ---
 title: Directory Structure
 desc: (@quasar/app-vite) The structure of a Quasar app with explanations for each folder and file.
+components:
+  - directory-structure/app-vite/DirectoryStructure
 ---
 This is the structure of a project with all modes installed. There's no reason to be intimidated though!
 
@@ -8,4 +10,4 @@ This is the structure of a project with all modes installed. There's no reason t
 If you are a beginner, all you'll need to care about is `/quasar.config.js` (Quasar App Config file), `/src/router`, `/src/layouts`, `/src/pages` and optionally `/src/assets`.
 :::
 
-<doc-q-component file="QuasarCliVite/DirectoryStructure/DirectoryStructure" />
+<directory-structure />
